requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
nltk==3.9.1
pymongo==4.13.2
python-dotenv==1.1.1
regex==2024.11.6
tqdm==4.67.1
Werkzeug==3.1.3

requests~=2.32.4